requirements.txt
numpy==1.23.5
gym==0.21.0
direct==1.0.1
panda3d==1.10.13
pandas==1.5.2
pygame==2.1.0
PyYAML==6.0
stable_baselines3==1.6.2
